test: share DBF memo fixture helpers
