+++ CHANGELOG.md
@@ -28,6 +28,10 @@
   lifetime, fault-containment, and owned-Windows-verification boundaries. No
    COM-event runtime behavior is implemented or claimed yet.
 
+- 2026-08-23: Consolidated shared DBF binary-fixture helpers in the focused
+  DBF-table test support module, removing duplicate helper implementations
+   from the memo-integrity test split without changing DBF behavior.
+
 - 2026-08-23: Continued #2564 test-module refactoring by moving the basic DBF
   memo and G/P create/replace/append lifecycle regressions into
   `test_dbf_table_memo_lifecycle.cpp`. The existing `test_dbf_table`
