+++ tests/test_dbf_table.cpp
@@ -42,6 +42,16 @@ namespace {
 using copperfin::test_support::ScopedEnvironmentValue;
 using copperfin::test_dbf_table::expect;
 using copperfin::test_dbf_table::failures;
+using copperfin::test_dbf_table::read_be_u16;
+using copperfin::test_dbf_table::read_binary_file;
+using copperfin::test_dbf_table::replace_memo_block_payload;
+using copperfin::test_dbf_table::write_ascii;
+using copperfin::test_dbf_table::write_be_u16;
+using copperfin::test_dbf_table::write_be_u32;
+using copperfin::test_dbf_table::write_binary_file;
+using copperfin::test_dbf_table::write_field_descriptor;
+using copperfin::test_dbf_table::write_le_u16;
+using copperfin::test_dbf_table::write_le_u32;
 
 class comma_decimal_numpunct final : public std::numpunct<char> {
 protected:
@@ -82,65 +92,6 @@ std::size_t count_missing_locale_keys(
     return missing;
 }
 
-void write_le_u16(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint16_t value) {
-    bytes[offset] = static_cast<std::uint8_t>(value & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-}
-
-void write_le_u32(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint32_t value) {
-    bytes[offset] = static_cast<std::uint8_t>(value & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-    bytes[offset + 2U] = static_cast<std::uint8_t>((value >> 16U) & 0xFFU);
-    bytes[offset + 3U] = static_cast<std::uint8_t>((value >> 24U) & 0xFFU);
-}
-
-void write_be_u16(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint16_t value) {
-    bytes[offset] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>(value & 0xFFU);
-}
-
-void write_be_u32(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint32_t value) {
-    bytes[offset] = static_cast<std::uint8_t>((value >> 24U) & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>((value >> 16U) & 0xFFU);
-    bytes[offset + 2U] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-    bytes[offset + 3U] = static_cast<std::uint8_t>(value & 0xFFU);
-}
-
-void write_ascii(std::vector<std::uint8_t>& bytes, std::size_t offset, const std::string& value) {
-    for (std::size_t index = 0; index < value.size(); ++index) {
-        bytes[offset + index] = static_cast<std::uint8_t>(value[index]);
-    }
-}
-
-void write_field_descriptor(
-    std::vector<std::uint8_t>& bytes,
-    std::size_t offset,
-    const std::string& name,
-    char type,
-    std::uint32_t field_offset,
-    std::uint8_t field_length,
-    std::uint8_t decimal_count = 0U) {
-    write_ascii(bytes, offset, name);
-    bytes[offset + 11U] = static_cast<std::uint8_t>(type);
-    write_le_u32(bytes, offset + 12U, field_offset);
-    bytes[offset + 16U] = field_length;
-    bytes[offset + 17U] = decimal_count;
-}
-
-std::vector<std::uint8_t> read_binary_file(const std::filesystem::path& path) {
-    std::ifstream input(path, std::ios::binary);
-    return {
-        std::istreambuf_iterator<char>(input),
-        std::istreambuf_iterator<char>()
-    };
-}
-
-bool write_binary_file(const std::filesystem::path& path, const std::vector<std::uint8_t>& bytes) {
-    std::ofstream output(path, std::ios::binary | std::ios::trunc);
-    output.write(reinterpret_cast<const char*>(bytes.data()), static_cast<std::streamsize>(bytes.size()));
-    return static_cast<bool>(output);
-}
-
 void clear_dbf_last_update_date(const std::filesystem::path& path) {
     std::vector<std::uint8_t> bytes = read_binary_file(path);
     expect(bytes.size() >= 4U, "DBF date-stamping fixture should contain a complete header");
@@ -161,39 +112,6 @@ void expect_dbf_last_update_date(const std::filesystem::path& path, const std::s
                    result.header.last_update_day >= 1U && result.header.last_update_day <= 31U,
                label + " should stamp a valid DBF last-update header date");
     }
-}
-
-std::uint16_t read_be_u16(const std::vector<std::uint8_t>& bytes, std::size_t offset) {
-    return static_cast<std::uint16_t>(
-        (static_cast<std::uint16_t>(bytes[offset]) << 8U) |
-        static_cast<std::uint16_t>(bytes[offset + 1U]));
-}
-
-bool replace_memo_block_payload(
-    const std::filesystem::path& memo_path,
-    std::uint32_t block_number,
-    const std::vector<std::uint8_t>& payload) {
-    std::vector<std::uint8_t> bytes = read_binary_file(memo_path);
-    if (bytes.size() < 8U || block_number == 0U) {
-        return false;
-    }
-    const std::uint16_t block_size = read_be_u16(bytes, 6U);
-    const std::size_t block_offset = static_cast<std::size_t>(block_number) * block_size;
-    if (block_size < 8U || block_offset + block_size > bytes.size() ||
-        payload.size() > static_cast<std::size_t>(block_size - 8U)) {
-        return false;
-    }
-
-    write_be_u32(bytes, block_offset + 4U, static_cast<std::uint32_t>(payload.size()));
-    std::fill(
-        bytes.begin() + static_cast<std::ptrdiff_t>(block_offset + 8U),
-        bytes.begin() + static_cast<std::ptrdiff_t>(block_offset + block_size),
-        static_cast<std::uint8_t>(0U));
-    std::copy(
-        payload.begin(),
-        payload.end(),
-        bytes.begin() + static_cast<std::ptrdiff_t>(block_offset + 8U));
-    return write_binary_file(memo_path, bytes);
 }
 
 std::vector<std::uint8_t> read_record_field_bytes(
