+++ tests/test_dbf_table_memo_integrity.cpp
@@ -6,12 +6,8 @@
 
 #include "copperfin/vfp/dbf_table.h"
 
-#include <algorithm>
-#include <cstddef>
-#include <cstdint>
 #include <filesystem>
 #include <fstream>
-#include <iterator>
 #include <string>
 #include <vector>
 
@@ -23,94 +19,6 @@
 #endif
 
 namespace copperfin::test_dbf_table {
-namespace {
-
-void write_le_u16(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint16_t value) {
-    bytes[offset] = static_cast<std::uint8_t>(value & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-}
-
-void write_le_u32(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint32_t value) {
-    bytes[offset] = static_cast<std::uint8_t>(value & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-    bytes[offset + 2U] = static_cast<std::uint8_t>((value >> 16U) & 0xFFU);
-    bytes[offset + 3U] = static_cast<std::uint8_t>((value >> 24U) & 0xFFU);
-}
-
-void write_be_u16(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint16_t value) {
-    bytes[offset] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>(value & 0xFFU);
-}
-
-void write_be_u32(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint32_t value) {
-    bytes[offset] = static_cast<std::uint8_t>((value >> 24U) & 0xFFU);
-    bytes[offset + 1U] = static_cast<std::uint8_t>((value >> 16U) & 0xFFU);
-    bytes[offset + 2U] = static_cast<std::uint8_t>((value >> 8U) & 0xFFU);
-    bytes[offset + 3U] = static_cast<std::uint8_t>(value & 0xFFU);
-}
-
-void write_ascii(std::vector<std::uint8_t>& bytes, std::size_t offset, const std::string& value) {
-    for (std::size_t index = 0; index < value.size(); ++index) {
-        bytes[offset + index] = static_cast<std::uint8_t>(value[index]);
-    }
-}
-
-void write_field_descriptor(
-    std::vector<std::uint8_t>& bytes,
-    std::size_t offset,
-    const std::string& name,
-    char type,
-    std::uint32_t field_offset,
-    std::uint8_t field_length) {
-    write_ascii(bytes, offset, name);
-    bytes[offset + 11U] = static_cast<std::uint8_t>(type);
-    write_le_u32(bytes, offset + 12U, field_offset);
-    bytes[offset + 16U] = field_length;
-}
-
-std::vector<std::uint8_t> read_binary_file(const std::filesystem::path& path) {
-    std::ifstream input(path, std::ios::binary);
-    return {std::istreambuf_iterator<char>(input), std::istreambuf_iterator<char>()};
-}
-
-bool write_binary_file(const std::filesystem::path& path, const std::vector<std::uint8_t>& bytes) {
-    std::ofstream output(path, std::ios::binary | std::ios::trunc);
-    output.write(reinterpret_cast<const char*>(bytes.data()), static_cast<std::streamsize>(bytes.size()));
-    return static_cast<bool>(output);
-}
-
-std::uint16_t read_be_u16(const std::vector<std::uint8_t>& bytes, std::size_t offset) {
-    return static_cast<std::uint16_t>(
-        (static_cast<std::uint16_t>(bytes[offset]) << 8U) |
-        static_cast<std::uint16_t>(bytes[offset + 1U]));
-}
-
-bool replace_memo_block_payload(
-    const std::filesystem::path& memo_path,
-    std::uint32_t block_number,
-    const std::vector<std::uint8_t>& payload) {
-    std::vector<std::uint8_t> bytes = read_binary_file(memo_path);
-    if (bytes.size() < 8U || block_number == 0U) {
-        return false;
-    }
-    const std::uint16_t block_size = read_be_u16(bytes, 6U);
-    const std::size_t block_offset = static_cast<std::size_t>(block_number) * block_size;
-    if (block_size < 8U || block_offset + block_size > bytes.size() ||
-        payload.size() > static_cast<std::size_t>(block_size - 8U)) {
-        return false;
-    }
-
-    write_be_u32(bytes, block_offset + 4U, static_cast<std::uint32_t>(payload.size()));
-    std::fill(bytes.begin() + static_cast<std::ptrdiff_t>(block_offset + 8U),
-              bytes.begin() + static_cast<std::ptrdiff_t>(block_offset + block_size),
-              static_cast<std::uint8_t>(0U));
-    std::copy(payload.begin(), payload.end(),
-              bytes.begin() + static_cast<std::ptrdiff_t>(block_offset + 8U));
-    return write_binary_file(memo_path, bytes);
-}
-
-}  // namespace
-
 void test_pack_memo_fails_when_referenced_payload_cannot_be_recovered() {
     namespace fs = std::filesystem;
     const fs::path temp_dir = fs::temp_directory_path() /
