+++ tests/test_dbf_table_support.h
@@ -4,13 +4,38 @@
 
 #pragma once
 
+#include <cstddef>
+#include <cstdint>
+#include <filesystem>
 #include <string>
+#include <vector>
 
 namespace copperfin::test_dbf_table {
 
 extern int failures;
 
 void expect(bool condition, const std::string& message);
+
+void write_le_u16(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint16_t value);
+void write_le_u32(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint32_t value);
+void write_be_u16(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint16_t value);
+void write_be_u32(std::vector<std::uint8_t>& bytes, std::size_t offset, std::uint32_t value);
+void write_ascii(std::vector<std::uint8_t>& bytes, std::size_t offset, const std::string& value);
+void write_field_descriptor(
+    std::vector<std::uint8_t>& bytes,
+    std::size_t offset,
+    const std::string& name,
+    char type,
+    std::uint32_t field_offset,
+    std::uint8_t field_length,
+    std::uint8_t decimal_count = 0U);
+std::vector<std::uint8_t> read_binary_file(const std::filesystem::path& path);
+bool write_binary_file(const std::filesystem::path& path, const std::vector<std::uint8_t>& bytes);
+std::uint16_t read_be_u16(const std::vector<std::uint8_t>& bytes, std::size_t offset);
+bool replace_memo_block_payload(
+    const std::filesystem::path& memo_path,
+    std::uint32_t block_number,
+    const std::vector<std::uint8_t>& payload);
 
 void test_memo_field_create_replace_and_append_round_trip();
 void test_general_and_picture_memo_fields_round_trip();
